pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1",
    "djangorestframework>=3.16.1",
]

[dependency-groups]
dev = [
    "pytest-django>=4.11.1",
    "ruff>=0.14.10",
]

[pytest]
DJANGO_SETTINGS_MODULE = "conf.settings"
python_files = ["test_*.py"]